+++ NEWS.md
@@ -2,6 +2,8 @@
 -----
 * Fix skipfields use with SHP output format (#81)
 * Fix Content-Disposition for error responses (#82)
+* Add pid to /cachestatus (#83)
+* Check CDB_QueryTable response before saving into cache (#83)
 
 1.3.4 (21/01/13)
 -----